+++ modules/apps/prompt.js
@@ -18,10 +18,13 @@ export class HMPrompt extends Application {
         this.dialogData = dialogData;
     }
 
-    getCapList(capsArr, actor=null) {
+    getCapList(weapon, actor=null) {
+        const capsArr = weapon.capabilities;
         const {special} = idx;
         const capsObj = Object.fromEntries(capsArr.map((x) => Object.entries(special)[x]));
-        actor.canBackstab
+        const ranged = weapon.data.data.ranged.checked;
+
+        (actor.canBackstab && !ranged)
             ? capsObj[HMCONST.SPECIAL.FLEEING] = special[HMCONST.SPECIAL.FLEEING]
             : delete capsObj[HMCONST.SPECIAL.BACKSTAB];
         return capsObj;